requirements.txt
python-telegram-bot==20.7
google-generativeai==0.8.3
python-docx==1.1.2
python-dotenv==1.0.0
